kernel

+++ preprocessing/blur_utils.py
@@ -12,12 +12,12 @@ def _gaussian_kernel(kernel_size, sigma, n_channels, dtype):
     return tf.expand_dims(tf.tile(g_kernel, (1, 1, n_channels)), axis=-1)
 
 def _x_motion_kernel(kernel_size):
-    np_result = np.zeros((kernel_size, kernel_size)).astype(np.float)
+    np_result = np.zeros((kernel_size, kernel_size)).astype(np.float32)
     np_result[:, kernel_size//2] = 1
     return tf.convert_to_tensor(np_result)
 
 def _y_motion_kernel(kernel_size):
-    np_result = np.zeros((kernel_size, kernel_size)).astype(np.float)
+    np_result = np.zeros((kernel_size, kernel_size)).astype(np.float32)
     np_result[kernel_size//2, :] = 1
     return tf.convert_to_tensor(np_result)
